backend/requirements.txt
# Essential packages
fastapi>=0.95.0,<1.0.0
uvicorn>=0.22.0,<0.23.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# OpenAI
openai>=1.0.0
tiktoken>=0.4.0

# Utilities
requests>=2.28.0
numpy>=1.22.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
httpx>=0.23.0

# Web and API
aiohttp>=3.8.0

# Security
cryptography>=36.0.0
bcrypt>=4.0.0

# Testing
pytest>=7.0.0

# Monitoring (optional)
prometheus-client>=0.16.0
sentry-sdk>=1.14.0

# Dependencies for pydantic-settings
pydantic-settings==2.0.3
typing-extensions==4.7.1